+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Added `delete-account` alias for `delete-user` command.
+- Support loading featured posts using `/api/v1/accounts/{account_id}/load_activities` API endpoint.
 
 ### Changed